move all the data handling to the newTaskFrom component
